Quantity update
SID update.
Now if a item share the same and the same date as an existing item, a
quantity counter is updated insetead of creating a new entry.
The dictionnary is now represented as follow:
NAME: [DATE, QUANTITY].

Reading, storing and exporting are OK but need further testing.
Adding and display are NOT updated yet.

Several code syntax errors fixed.